Added different components and added a function to compareDates
